Add basic authorization

+++ lib/croods/controller/actions.rb
@@ -4,23 +4,29 @@ module Croods
   class Controller < ActionController::API
     module Actions
       def index
-        render json: model.order(:created_at)
+        authorize model
+        render json: policy_scope(model).order(:created_at)
       end
 
       def show
+        authorize member
         render json: member
       end
 
       def create
-        render status: :created, json: model.create!(member_params)
+        authorize model
+        render status: :created, json: policy_scope(model)
+          .create!(member_params)
       end
 
       def update
+        authorize member
         member.update!(member_params)
         render json: member
       end
 
       def destroy
+        authorize member
         render json: member.destroy!
       end
     end
